requirement.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
twilio==7.0.0
sendgrid==6.9.7
